+++ src/main/java/br/edu/utfpr/chemistsincontrol/service/EquipamentoService.java
@@ -2,4 +2,4 @@ package br.edu.utfpr.chemistsincontrol.service;
 
 import br.edu.utfpr.chemistsincontrol.model.Equipamento;
 
-public interface EquipamentoService extends CrudService<Equipamento, Long>{}
+public interface EquipamentoService extends CrudService<Equipamento, Integer>{}
